requirements.txt
absl-py==0.10.0
astor==0.8.1
certifi==2020.6.20
chardet==3.0.4
dlib==19.21.0
facenet-pytorch==2.4.1
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.19.2
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
requests==2.24.0
scipy==1.5.2
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
torch==1.6.0
torchvision==0.7.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.2.0
